% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album.R
\name{album}
\alias{album}
\title{Generate album color palette}
\usage{
album(n, alpha = 1, begin = 0, end = 1, direction = 1, lp = "in_rainbows")
}
\arguments{
\item{lp}{}
}
\value{
a color palette
}
\description{
Generate album color palette
}
