% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/album_pal.R
\name{album_pal}
\alias{album_pal}
\title{Generate album color palette}
\usage{
album_pal(n, alpha = 1, begin = 0, end = 1, direction = 1, lp = "in_rainbows")
}
\arguments{
\item{alpha}{pass through parameter to ggplot}

\item{begin}{The (corrected) hue in \link{0,1} at which the radiohead colormap begins.}

\item{end}{The (corrected) hue in \link{0,1} at which the radiohead colormap begins.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are as output by album_pal If -1, the order of colors is reversed.}

\item{lp}{the album to select, e.g. lp = "ok_computer"}
}
\value{
a color palette
}
\description{
Generate album color palette
}
