% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-palette.R
\name{plot_palette}
\alias{plot_palette}
\title{Plot Palette}
\usage{
plot_palette(lp)
}
\arguments{
\item{lp}{the album to plot}
}
\value{
a ggplot2 object
}
\description{
Plots the color palette for the album in question
}
\examples{

plot_palette(lp = "kid_a")

}
