% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color-radiohead.R
\name{scale_color_radiohead}
\alias{scale_color_radiohead}
\title{Scale color radiohead}
\usage{
scale_color_radiohead(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  album = "in_rainbows",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "color",
  discrete = TRUE
)
}
\arguments{
\item{discrete}{}
}
\value{

}
\description{
Scale color radiohead
}
