% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color-radiohead.R
\name{scale_color_radiohead_c}
\alias{scale_color_radiohead_c}
\title{scale_color_radiohead_c}
\usage{
scale_color_radiohead_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  album = "in_rainbows",
  n = 256
)
}
\arguments{
\item{alpha}{pass through parameter to ggplot}

\item{begin}{The (corrected) hue in \link{0,1} at which the radiohead colormap begins.}

\item{end}{The (corrected) hue in \link{0,1} at which the radiohead colormap begins.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are as output by album_pal If -1, the order of colors is reversed.}

\item{album}{the radiohead album to select, e.g. album = "ok_computer"}

\item{aesthetics}{}
}
\value{

}
\description{
Create a continuous color scale based on a selected radiohead album
}
