% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-fill-album.R
\name{scale_fill_album_c}
\alias{scale_fill_album_c}
\title{scale_fill_album_c}
\usage{
scale_fill_album_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  album = "in_rainbows",
  n = 256
)
}
\arguments{
\item{alpha}{pass through parameter to ggplot}

\item{begin}{The (corrected) hue in \link{0,1} at which the radiohead colormap begins.}

\item{end}{The (corrected) hue in \link{0,1} at which the radiohead colormap begins.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are as output by album_pal If -1, the order of colors is reversed.}

\item{album}{the album to select, e.g. album = "ok_computer"}

\item{aesthetics}{}
}
\value{

}
\description{
#' Create a continuous fill based on an album
}
