% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-fill-radiohead.R
\name{scale_fill_radiohead_c}
\alias{scale_fill_radiohead_c}
\title{scale_fill_radiohead_c}
\usage{
scale_fill_radiohead_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  album = "in_rainbows",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill",
  discrete = TRUE
)
}
\arguments{
\item{discrete}{}
}
\value{

}
\description{
scale_fill_radiohead_c
}
