% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-fill-radiohead.R
\name{scale_fill_radiohead_d}
\alias{scale_fill_radiohead_d}
\title{scale_fill_radiohead_d
Uses the radiohead album selected to create a discrete fill}
\usage{
scale_fill_radiohead_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  album = "in_rainbows",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill",
  discrete = TRUE
)
}
\arguments{
\item{alpha}{pass through parameter to ggplot}

\item{begin}{The (corrected) hue in \link{0,1} at which the radiohead colormap begins.}

\item{end}{The (corrected) hue in \link{0,1} at which the radiohead colormap begins.}

\item{direction}{Sets the order of colors in the scale. If 1, the default, colors are as output by viridis_pal. If -1, the order of colors is reversed.}

\item{album}{the radiohead album to select, e.g. album = "ok_computer"}

\item{discrete}{}
}
\value{

}
\description{
scale_fill_radiohead_d
Uses the radiohead album selected to create a discrete fill
}
