% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggroove-pal.R
\name{ggroove_pal}
\alias{ggroove_pal}
\title{used for discrete color palette}
\usage{
ggroove_pal(alpha = 1, begin = 0, end = 1, direction = 1, lp = "in_rainbows")
}
\arguments{
\item{lp}{}
}
\value{
discrete palette
}
\description{
used for discrete color palette
}
