% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-color-radiohead.R
\name{scale_color_radiohead_c}
\alias{scale_color_radiohead_c}
\title{scale_color_radiohead_c}
\usage{
scale_color_radiohead_c(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  album = "in_rainbows",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "color",
  discrete = TRUE
)
}
\arguments{
\item{album}{the radiohead album to select, e.g. album = "ok_computer"}

\item{discrete}{}
}
\value{

}
\description{
Create a continuous color scale based on a selected radiohead album
}
